Consolidation of duplicate code into spoonacular.js
