+++ source/assets/components/recipeCard.js
@@ -1,36 +1,36 @@
 //recipeCard.js
-
-/*********  CARD STRUCTURE ***********
-  <recipe-card>
-  ***** Shadow DOM *****
-    <section>
-      <main id='card-information'>
-        <div id='recipe-title'> </div>
-        <div id='recipe-price'> </div>
-        <div id='recipe-cooking-time'> </div>
-        <div id='recipe-score'> </div>
-        <div id='recipe-servings'> </div>
-        <div id='recipe-calories'> </div>
-        <div id='recipe-total-calories'> </div>
-      </main>
-      <aside id ='card-visuals'>
-        <fig id='visual'>
-          <img id='recipe-image'/>
-          <figcaption id='recipe-dietary'> </figcaption>
-        </fig>
-      </aside>
-    </section> 
-  ***** END OF SHADOW DOM *****
-  <recipe-card>
-***************************************/
+import { Spoonacular } from "../scripts/spoonacular.js";
 
 class recipeCard extends HTMLElement {
   constructor() {
     super();
     this.attachShadow({ mode: "open" });
   }
-
+  /*********  CARD STRUCTURE ***********
+    <recipe-card>
+    ***** Shadow DOM *****
+      <section>
+        <main id='card-information'>
+          <div id='recipe-title'> </div>
+          <div id='recipe-price'> </div>
+          <div id='recipe-cooking-time'> </div>
+          <div id='recipe-score'> </div>
+          <div id='recipe-servings'> </div>
+          <div id='recipe-calories'> </div>
+          <div id='recipe-total-calories'> </div>
+        </main>
+        <aside id ='card-visuals'>
+          <fig id='visual'>
+            <img id='recipe-image'/>
+            <figcaption id='recipe-dietary'> </figcaption>
+          </fig>
+        </aside>
+      </section> 
+    ***** END OF SHADOW DOM *****
+    <recipe-card>
+  ***************************************/
   set data(data) {
+    const spoonacular = new Spoonacular();
     const card = document.createElement("section");
     const style = document.createElement("style");
     style.innerHTML = `
@@ -94,14 +94,14 @@ class recipeCard extends HTMLElement {
       /** RECIPE IMAGE */
       const recipeImage = document.createElement("img");
       recipeImage.id = "recipe-image";
-      recipeImage.src = getRecipeImageSource(data);
+      recipeImage.src = spoonacular.getRecipeImageSource(data);
 
       /** RECIPE IMAGE CAPTION */
       const recipeDietary = document.createElement("figcaption");
       recipeDietary.id = "recipe-dietary";
       
       function setRecipeDietary(){
-        const dietary = getRecipeDietary(data);
+        const dietary = spoonacular.getRecipeDietary(data);
         let dietaryIcon;
         if(dietary["vegan"]){
           dietaryIcon = document.createElement('img');
@@ -140,37 +140,37 @@ class recipeCard extends HTMLElement {
       /** RECIPE TITLE */
       const recipeTitle = document.createElement("div");
       recipeTitle.id = "recipe-title";
-      recipeTitle.innerHTML = getRecipeTitle(data);
+      recipeTitle.innerHTML = spoonacular.getRecipeTitle(data);
 
       /** RECIPE PRICE */
       const recipePrice = document.createElement("div");
       recipePrice.id = "recipe-price";
-      recipePrice.innerHTML = getRecipePrice(data);
+      recipePrice.innerHTML = spoonacular.getRecipePrice(data);
 
       /** RECIPE COOKING TIME */
       const recipeCookingTime = document.createElement("div");
       recipeCookingTime.id = "recipe-cooking-time";
-      recipeCookingTime.innerHTML = getRecipeCookingTime(data);
+      recipeCookingTime.innerHTML = spoonacular.getRecipeCookingTime(data);
 
       /** RECIPE SCORE */
       const recipeScore = document.createElement("div");
       recipeScore.id = "recipe-score";
-      recipeScore.innerHTML = getRecipeScore(data);
+      recipeScore.innerHTML = spoonacular.getRecipeScore(data);
 
       /** RECIPE SERVING */
       const recipeServings = document.createElement("div");
       recipeServings.id = "recipe-servings";
-      recipeServings.innerHTML = getRecipeServings(data);
+      recipeServings.innerHTML = spoonacular.getRecipeServings(data);
 
       /** RECIPE CALORIES */
       const recipeCalories = document.createElement("div");
       recipeCalories.id = "recipe-calories";
-      recipeCalories.innerHTML = getRecipeCalories(data);
+      recipeCalories.innerHTML = spoonacular.getRecipeCalories(data);
 
       /** RECIPE TOTAL CALORIES */
       const recipeTotalCalories = document.createElement("div");
       recipeTotalCalories.id = "recipe-total-calories";
-      recipeTotalCalories.innerHTML = getRecipeTotalCalories(data);
+      recipeTotalCalories.innerHTML = spoonacular.getRecipeTotalCalories(data);
 
       cardInformation.appendChild(recipeTitle);
       cardInformation.appendChild(recipePrice);
@@ -190,75 +190,6 @@ class recipeCard extends HTMLElement {
     this.shadowRoot.appendChild(style);
     this.shadowRoot.appendChild(card);
   }
-}
-
-/** HELPER METHODS */
-function getRecipeImageSource(data) {
-  return data["image"];
-}
-
-function getRecipeTitle(data) {
-  return data["title"];
-}
-/**
- * function return the toal cost of the recipe
- * Servings * pricePerServings
- **/
-function getRecipePrice(data) {
-  return Math.round(data["pricePerServing"] * data["servings"]);
-}
-
-function getRecipeCookingTime(data) {
-  if (data) {
-    return "Minutes: " + data["readyInMinutes"];
-  }
-  return null;
-}
-
-function getRecipeScore(data) {
-  if (data) {
-    return "Score: " + data["spoonacularScore"] + "/100";
-  }
-  return null;
-}
-
-function getRecipeCalories(data) {
-  if (data) {
-    return (
-      "Calories per Serving: " +
-      Math.round(data["nutrition"]["nutrients"][0]["amount"])
-    );
-  }
-  return null;
-}
-
-function getRecipeServings(data) {
-  if (data) {
-    return "Servings: " + data["servings"];
-  }
-  return null;
-}
-
-function getRecipeTotalCalories(data) {
-  if (data) {
-    return (
-      "Total Calories: " +
-      Math.round(data["nutrition"]["nutrients"][0]["amount"] * data["servings"])
-    );
-  }
-  return null;
-}
-
-function getRecipeDietary(data){
-  if(data){
-    let dietary = {};
-    dietary["vegan"] = data["vegan"];
-    dietary["vegetarian"] = data["vegetarian"];
-    dietary["dairy-free"] = data["dairyFree"]; 
-    dietary["gluten-free"] = data["glutenFree"];
-    return dietary;
-  }
-  return null;
 }
 
 customElements.define("recipe-card", recipeCard);
