+++ source/assets/test_stuff/apiTest.html
@@ -7,6 +7,8 @@
     <title>Some Testing</title>
 
     <script src="apiTest.js" type="module"></script>
+    <!-- Spoonacular-->
+    <script src="../scripts/spoonacular.js" type="module"></script>
     <!-- Web Components -->
     <script src="../components/recipeCard.js" type="module"></script>
     <script src="../components/recipeViewer.js" type="module"></script>
