+++ source/index.html
@@ -3,6 +3,8 @@
 <head>
   <title>Recipe App</title>
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
+  <!-- Spoonacular -->
+  <script src="assets/scripts/spoonacular.js" type="module"></script>
   <!-- Components -->
   <script src="assets/components/recipeCard.js" type="module"></script>
   <script src="assets/components/recipeViewer.js" type="module"></script>
